requirements.txt
sqlalchemy
requests
pandas
python-dotenv